Fortran: generate tags for interface by default
Fortran interface is used to define generic subprograms, wrap subprogram
prototypes and overload operators. The tags for interfaces should be
generated for the first and third case for that interfaces defines new
subprograms in these cases. For the second case, the interfaces can be
anonymous. If anonymous, tag for it will not be generated as expected.