C++: add imported role to member kind
This new role is for tagging X in

     using X;

as a reference tag.

What we can say from the statement is that an item is defined/declared
somewhere and it can be referred as X in the input source file.. We
cannot say the kind of the item from the input source file.

Temporary I choose `member' as a kind for the object.
This change records the fact that the member is "imported" as
the name X.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>