jscript: Improve and fix scope handling
Handle scope more consistently, avoiding dealing with it in many places
in a fragile and sometimes incorrect manner.

This fixes most scopes in current test cases.  Handling of qualified
names is still questionable, but mostly follows the current handling to
prevent breakage.  Better handling of these would require tracking
which tag the qualified name actually refers to to be able to choose
the correct scope, but that's outside the scope of this patch.