EXPERIMENTAL: run ctags process in parallel
   $ ./ctags -R --_parallel ~/var/linux

8 processes are launched as worker.
The main process provides input file name to the workers.
"consumer" process gathers output form the workers and write all to the stdout.
Sort should be done in the post processor like:

    $ ./ctags -R --_parallel ~/var/linux | sort > /tmp/P;echo; time  ./ctags -R --sort=no ~/var/linux | sort > /tmp/S;echo; time  ./ctags -R ~/var/linux > /tmp/S-builtin-sort; echo

"8" is hard-coded. --_parallel should take the number of workers.

The unit of consuming the output from workers done by "consuming" is LINE.
This makes the "consumer" CPU hogging.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>