DocBook: new parser imported from Geany
This pull request adds one of the extra parsers we have in Geany except
those defined in c.c and except the markdown parser which is readLine()
based and not really superior to the regex-based parser ctags uses
(we should probably use the ctags parser in Geany instead).

The added a parser is for DocBook (https://en.wikipedia.org/wiki/DocBook).

The original version is written for Geany by Enrico Tröger <enrico.troeger@uvena.de>:

    commit 46690c0b8f1e5b47bcbd4a2732233e539cfbf7c4
    Author: Enrico Tröger <enrico.troeger@uvena.de>
    Date:   Wed Jan 11 18:50:44 2006 +0000

	added tagmanager support for filetype DocBook and LaTeX

	git-svn-id: https://geany.svn.sourceforge.net/svnroot/geany/trunk@123 ea778897-0a13-0410-b9d1-a72fbfd435f5

This log message is edited by @masatake.
After getting the permission from the original author at #2144,
@masatake wrote the version of GPL at the head of the source code.