JavaScript: make tags for kinds even if they are disabled
This change is preparation for switching the JavaScript parser to Cork
based.

In cork based parser, a name (A) being part of a scope of another name
(B) must be recorded to the cork queue to make B's scope string even if
the kind for A is disabled.

In stead of the JavsScript parser, the main part controls whether
ctags writes the tag for A to tags file.

"codebase ctags JavaScript" doesn't report any significant performance
degradation for this change.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>