main: pass a line to lregex layer after chopping \n
The ctags main part passed a input line to a single-line regex parser
before chopping \n. This was o.k. when ctags used only posix regex engine.

We are planning to introduce another regex engine like pcre2.
It seems that the ways to handle \n in the pcre2 and posix regex engine
are different. See https://www.pcre.org/current/doc/html/pcre2posix.html.

To minimize the impact of the difference, this change hides the newline
in a input line when passing the line to the regex engine.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>