main,refactor: use a loop for printing extension fields in tags output
The fields are hard-coded in the ctags writer.
This makes adding a new field a bit harder.

This change refactors the way to print extension fields; we use a loop
instead of the hard-conding.

This commit changes the order of key-value pairs in the json output.
We need an ack from https://github.com/liuchengxu/vista.vim before merging.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>